Rename DOS county page submodules.
